Various improvements to buffer shuffling algorithm
Now only output whole buffers (except for the final chunk of course),
so we get consistent large chunks. Use 32k output chunks for
decompression and 16k chunks for compression.
Be more particular with the invariants and the possability of an empty 
input buffer. That is now impossible for compression and indicates an 
error condition for decompression.
Tidy up invariant assertions.
Remove tracing & debug code.