Change internals to return data errors explicitly
Based on an original patch by Ganesh Sittampalam <ganesh@earth.li>
The point is that callers that are interested in the errors can now
see and manipulate them as values rather than just getting calls to
error. In particular this should allow darcs to read old files that
have incorrect data checksums.