Change the window bits range from 8..15 to 9..15
The zlib C lib has been changed to clarify that in fact window bits = 8
is not supported:

https://github.com/madler/zlib/commit/049578f0a1849f502834167e233f4c1d52ddcbcc

We already had a guard in one of the tests to avoid the window bits = 8
case for some formats where it had been observed to fail. The change in
the C lib now makes it fail in more cases (since it had never been
properly supported anyway). So the simplest thing to do is just say that
9 is the minimum value.

Credit to Sergei Trofimovich for notifying and then tracking down the
root cause of the failure with newer versions of the C lib.

This fixes issue #11.